requirements.txt
# 🚀 Zero To Hero - Python Dependencies

# Django Framework
Django>=5.0,<6.0

# PostgreSQL Database Adapter
psycopg2-binary>=2.9.9

# Django Admin Interface Enhancement
django-admin-interface>=0.28.0
django-colorfield>=0.11.0

# Environment Variables
python-dotenv>=1.0.0

# Image Processing (for user avatars, platform logos)
Pillow>=10.0.0

# Payment Gateway
stripe>=7.0.0

# Email Validation
email-validator>=2.1.0

# Security
django-cors-headers>=4.3.0

# Production Server (optional - for deployment)
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0

# Development Tools (optional)
django-debug-toolbar>=4.2.0
Django==6.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
razorpay==2.0.0
requests==2.32.5
